+++ src/index.js
@@ -40,6 +40,8 @@ class Scheduler extends Component {
             contentHeight: schedulerData.getSchedulerContentDesiredHeight(),
             browserScrollbarHeight: 17,
             browserScrollbarWidth: 17,
+            scrollLeft: 0,
+            scrollTop: 0,
         };
     }
 
@@ -68,6 +70,10 @@ class Scheduler extends Component {
         slotClickedFunc: PropTypes.func,
         slotItemTemplateResolver: PropTypes.func,
         nonAgendaCellHeaderTemplateResolver: PropTypes.func,
+        onScrollLeft: PropTypes.func,
+        onScrollRight: PropTypes.func,
+        onScrollTop: PropTypes.func,
+        onScrollBottom: PropTypes.func,
     }
 
     componentDidMount(props, state){
@@ -346,6 +352,28 @@ class Scheduler extends Component {
             if (this.schedulerResource.scrollTop != this.schedulerContent.scrollTop)
                 this.schedulerResource.scrollTop = this.schedulerContent.scrollTop;
         }
+
+        const {schedulerData, onScrollLeft, onScrollRight, onScrollTop, onScrollBottom } = this.props;
+        const {scrollLeft, scrollTop} = this.state;
+        if(this.schedulerContent.scrollLeft !== scrollLeft) {
+            if(this.schedulerContent.scrollLeft === 0 && onScrollLeft != undefined) {
+                onScrollLeft(schedulerData, this.schedulerContent, this.schedulerContent.scrollWidth - this.schedulerContent.clientWidth);
+            }
+            if(this.schedulerContent.scrollLeft === this.schedulerContent.scrollWidth - this.schedulerContent.clientWidth && onScrollRight != undefined) {
+                onScrollRight(schedulerData, this.schedulerContent, this.schedulerContent.scrollWidth - this.schedulerContent.clientWidth);
+            }
+        } else if(this.schedulerContent.scrollTop !== scrollTop) {
+            if(this.schedulerContent.scrollTop === 0 && onScrollTop != undefined) {
+                onScrollTop(schedulerData, this.schedulerContent, this.schedulerContent.scrollHeight - this.schedulerContent.clientHeight);
+            }
+            if(this.schedulerContent.scrollTop === this.schedulerContent.scrollHeight - this.schedulerContent.clientHeight && onScrollBottom != undefined) {
+                onScrollBottom(schedulerData, this.schedulerContent, this.schedulerContent.scrollHeight - this.schedulerContent.clientHeight);
+            }
+        }
+        this.setState({
+            scrollLeft: this.schedulerContent.scrollLeft,
+            scrollTop: this.schedulerContent.scrollTop
+        });
     }
 
     onViewChange = (e) => {